backend/requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0
aiohttp==3.9.3

# Computer Vision & AI Models
torch==2.2.0
torchvision==0.17.0
opencv-python-headless==4.9.0.80
ultralytics==8.1.24
transformers==4.38.1
pillow==10.2.0
numpy==1.26.4

# Video Stream Processing
av==11.0.0
ffmpeg-python==0.2.0

# LLM & AI SDKs
anthropic==0.17.0
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# Security
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Email
aiosmtplib==3.0.1

# Redis (for frame queue)
redis==5.0.1
aioredis==2.0.1

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0
